requirements.txt
Django==1.10.4
dj-database-url==0.4.1
gunicorn==19.6.0
whitenoise==3.2.2
psycopg2==2.6.2
djangorestframework==3.6.3
